Small bug fixes

+++ controllers/documentController.mjs
@@ -1,4 +1,5 @@
 import Document from '../models/documentModel.mjs';
+import mongoose from 'mongoose';
 
 export async function getAll(req, res) {
     try {
@@ -6,19 +7,22 @@ export async function getAll(req, res) {
         res.status(200).json(docs);
     } catch (e) {
         res.status(500).json({
-            errors: {
-                status: 500,
-                type: 'get',
-                source: '/',
-                title: 'Database error',
-                detail: e.message,
-            },
+            status: 500,
+            type: 'get',
+            source: '/',
+            title: 'Database error',
+            detail: e.message,
         });
     }
 }
 
 export async function getOne(req, res) {
     const id = req.params.id;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ message: 'No doc found with that ID' });
+    }
+
     try {
         const doc = await Document.findById(id);
 
@@ -27,16 +31,15 @@ export async function getOne(req, res) {
                 .status(404)
                 .json({ message: 'No doc found with that ID' });
         }
+
         res.status(200).json(doc);
     } catch (e) {
         res.status(500).json({
-            errors: {
-                status: 500,
-                type: 'get',
-                source: '/:id',
-                title: 'Database error',
-                detail: e.message,
-            },
+            status: 500,
+            type: 'get',
+            source: '/:id',
+            title: 'Database error',
+            detail: e.message,
         });
     }
 }
@@ -56,13 +59,11 @@ export async function addOne(req, res) {
         res.status(201).json(doc);
     } catch (e) {
         res.status(500).json({
-            errors: {
-                status: 500,
-                type: 'post',
-                source: '/',
-                title: 'Database error',
-                detail: e.message,
-            },
+            status: 500,
+            type: 'post',
+            source: '/',
+            title: 'Database error',
+            detail: e.message,
         });
     }
 }
@@ -71,6 +72,10 @@ export async function updateOne(req, res) {
     const title = req.body.title;
     const content = req.body.content;
     const id = req.params.id;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ message: 'No doc found with that ID' });
+    }
 
     if (!title || !content) {
         return res
@@ -84,16 +89,21 @@ export async function updateOne(req, res) {
             { title: title, content: content },
             { new: true, runValidators: true }
         );
-        res.status(201).json(doc);
+
+        if (!doc) {
+            return res
+                .status(404)
+                .json({ message: 'No doc found with that ID' });
+        }
+
+        res.status(200).json(doc);
     } catch (e) {
         res.status(500).json({
-            errors: {
-                status: 500,
-                type: 'put',
-                source: '/:id',
-                title: 'Database error',
-                detail: e.message,
-            },
+            status: 500,
+            type: 'put',
+            source: '/:id',
+            title: 'Database error',
+            detail: e.message,
         });
     }
 }
@@ -101,20 +111,27 @@ export async function updateOne(req, res) {
 export async function deleteOne(req, res) {
     const id = req.params.id;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ message: 'No doc found with that ID' });
+    }
+
     try {
         const doc = await Document.findOneAndDelete({ _id: id });
-        res.status(200).json({
-            status: 'deleted',
-        });
+
+        if (!doc) {
+            return res
+                .status(404)
+                .json({ message: 'No doc found with that ID' });
+        }
+
+        res.status(200).json({ message: 'Document deleted successfully' });
     } catch (e) {
         res.status(500).json({
-            errors: {
-                status: 500,
-                type: 'delete',
-                source: '/:id',
-                title: 'Database error',
-                detail: e.message,
-            },
+            status: 500,
+            type: 'delete',
+            source: '/:id',
+            title: 'Database error',
+            detail: e.message,
         });
     }
 }
